special tests code refactored
